# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Portal Municipal Inteligente - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /noticias:
    get:
      operationId: listNoticias
      tags: [noticias]
      summary: Listar notícias
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: categoria
          in: query
          schema:
            type: string
        - name: destaque
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de notícias
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticiaListResponse"
    post:
      operationId: createNoticia
      tags: [noticias]
      summary: Criar notícia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticia"
      responses:
        "201":
          description: Notícia criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Noticia"
  /noticias/{slug}:
    get:
      operationId: getNoticiaBySlug
      tags: [noticias]
      summary: Obter notícia por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notícia encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Noticia"
        "404":
          description: Notícia não encontrada
    put:
      operationId: updateNoticia
      tags: [noticias]
      summary: Atualizar notícia
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticia"
      responses:
        "200":
          description: Notícia atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Noticia"
    delete:
      operationId: deleteNoticia
      tags: [noticias]
      summary: Remover notícia
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Notícia removida
...
  /servicos/{slug}:
    get:
      operationId: getServico
      tags: [servicos]
      summary: Obter serviço por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Serviço encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servico"
        "404":
          description: Serviço não encontrado
...
  /secretarias/{slug}:
    get:
      operationId: getSecretaria
      tags: [secretarias]
      summary: Obter secretaria por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Secretaria encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secretaria"
...
  /licitacoes:
    get:
      operationId: listLicitacoes
      tags: [licitacoes]
      summary: Listar licitações
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: modalidade
          in: query
          schema:
            type: string
        - name: situacao
          in: query
          schema:
            type: string
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de licitações
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicitacaoListResponse"
  /licitacoes/{id}:
    get:
      operationId: getLicitacao
      tags: [licitacoes]
      summary: Obter licitação por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Licitação encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Licitacao"
  /legislacao:
    get:
      operationId: listLegislacao
      tags: [legislacao]
      summary: Listar legislação municipal
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
            enum: [lei, decreto, portaria, resolucao, instrucao-normativa]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: ano
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de legislação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegislacaoListResponse"
  /legislacao/{tipo}/{slug}:
    get:
      operationId: getLegislacao
      tags: [legislacao]
      summary: Obter legislação por tipo e slug
      parameters:
        - name: tipo
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Legislação encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Legislacao"
...
  /busca:
    get:
      operationId: globalSearch
      tags: [busca]
      summary: Busca global no portal
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: tipo
          in: query
          schema:
            type: string
            enum: [noticia, servico, licitacao, legislacao, secretaria]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Resultados da busca
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuscaResponse"
...
components:
  schemas:
...
    Noticia:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        slug:
          type: string
        resumo:
          type: string
        conteudo:
          type: string
        imagemCapa:
          type: string
          nullable: true
        categoria:
          type: string
        autor:
          type: string
          nullable: true
        dataPublicacao:
          type: string
          format: date-time
        destaque:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - titulo
        - slug
        - resumo
        - conteudo
        - categoria
        - dataPublicacao
        - destaque
        - tags
    CreateNoticia:
      type: object
      properties:
        titulo:
          type: string
        slug:
          type: string
        resumo:
          type: string
        conteudo:
          type: string
        imagemCapa:
          type: string
          nullable: true
        categoria:
          type: string
        autor:
          type: string
          nullable: true
        dataPublicacao:
          type: string
          format: date-time
        destaque:
          type: boolean
          default: false
        tags:
          type: array
          items:
            type: string
      required:
        - titulo
        - slug
        - resumo
        - conteudo
        - categoria
        - dataPublicacao
    NoticiaListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Noticia"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Servico:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        slug:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        orgao:
          type: string
          nullable: true
        linkExterno:
          type: string
          nullable: true
        requisitos:
          type: array
          items:
            type: string
        documentos:
          type: array
          items:
            type: string
        prazoAtendimento:
          type: string
          nullable: true
        gratuito:
          type: boolean
        online:
          type: boolean
        icone:
          type: string
          nullable: true
      required:
        - id
        - titulo
        - slug
        - descricao
        - categoria
        - requisitos
        - documentos
        - gratuito
        - online
...
    Secretaria:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        slug:
          type: string
        sigla:
          type: string
        descricao:
          type: string
        secretario:
          type: string
          nullable: true
        fotoSecretario:
          type: string
          nullable: true
        telefone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        horario:
          type: string
          nullable: true
        competencias:
          type: array
          items:
            type: string
      required:
        - id
        - nome
        - slug
        - sigla
        - descricao
        - competencias
...
    Licitacao:
      type: object
      properties:
        id:
          type: string
        numero:
          type: string
        objeto:
          type: string
        modalidade:
          type: string
        situacao:
          type: string
        dataAbertura:
          type: string
          format: date-time
          nullable: true
        dataEncerrramento:
          type: string
          format: date-time
          nullable: true
        valorEstimado:
          type: number
          nullable: true
        valorHomologado:
          type: number
          nullable: true
        secretaria:
          type: string
          nullable: true
        edital:
          type: string
          nullable: true
        ata:
          type: string
          nullable: true
        descricao:
          type: string
          nullable: true
      required:
        - id
        - numero
        - objeto
        - modalidade
        - situacao
    LicitacaoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Licitacao"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Legislacao:
      type: object
      properties:
        id:
          type: string
        numero:
          type: string
        tipo:
          type: string
        ementa:
          type: string
        slug:
          type: string
        dataPublicacao:
          type: string
          format: date
        ano:
          type: integer
        conteudo:
          type: string
          nullable: true
        arquivoPdf:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - numero
        - tipo
        - ementa
        - slug
        - dataPublicacao
        - ano
        - tags
    LegislacaoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Legislacao"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...
    BuscaResultItem:
      type: object
      properties:
        id:
          type: string
        tipo:
          type: string
        titulo:
          type: string
        resumo:
          type: string
          nullable: true
        url:
          type: string
        data:
          type: string
          format: date-time
          nullable: true
        relevancia:
          type: number
      required:
        - id
        - tipo
        - titulo
        - url
        - relevancia
    BuscaResponse:
      type: object
      properties:
        query:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/BuscaResultItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - query
        - data
        - total
        - page
        - limit
        - totalPages